refactor: re-arrangement for gadgets for edit interfaces * merge PreviewWithVariant with PWV2017 * split AiAssisted * add patch to 2017 wikitext editor * optimize load logic
