+++ src/EditForm_JS/EditForm_JS.ts
@@ -1,9 +1,8 @@
 import {IS_WG_EDIT_OR_SUBMIT_ACTION, WG_NAMESPACE_NUMBER} from './modules/constant';
-import {aiAssisted} from './modules/aiAssisted';
 import {clearUndoSummary} from './modules/clearUndoSummary';
 import {disableTitle} from './modules/disableTitle';
 import {getBody} from 'ext.gadget.Util';
-import {introAFC} from './modules/introAFC';
+import {introACH} from './modules/introACH';
 import {preloadRevid} from './modules/preloadRevid';
 
 void getBody().then(function editForm($body: JQuery<HTMLBodyElement>): void {
@@ -22,19 +21,6 @@ void getBody().then(function editForm($body: JQuery<HTMLBodyElement>): void {
 	// 新用户引导至条目创建向导（[[QW:AFC]]）
 	const curid = mw.config.get('wgArticleId');
 	if (!curid && ![2, 3, 118].includes(WG_NAMESPACE_NUMBER)) {
-		introAFC();
+		introACH();
 	}
-
-	// AI辅助编辑特殊声明
-	mw.hook('wikipage.editform').add(($editForm): void => {
-		aiAssisted({
-			$editForm,
-		});
-	});
-
-	mw.hook('ve.saveDialog.stateChanged').add((): void => {
-		aiAssisted({
-			$body,
-		});
-	});
 });
