+++ src/Editform_AiAssisted/Editform_AiAssisted.ts
@@ -1,0 +1,21 @@
+import {getBody} from 'ext.gadget.Util';
+import {processVisualEditor} from './modules/processVisualEditor';
+import {processWikiEditor} from './modules/processWikiEditor';
+
+/**
+ * @description AI辅助编辑特殊声明
+ */
+void getBody().then(function editForm($body: JQuery<HTMLBodyElement>): void {
+	mw.hook('ve.saveDialog.stateChanged').add((): void => {
+		processVisualEditor({
+			$body,
+		});
+	});
+
+	mw.hook('wikipage.editform').add(($editForm): void => {
+		processWikiEditor({
+			$body,
+			$editForm,
+		});
+	});
+});
