+++ src/Editform_AiAssisted/definition.json
@@ -1,0 +1,8 @@
+{
+	"enable": true,
+	"description": "<span id=\"Gadget-Editform_AiAssisted\"></span> 声明编辑在人工智能（AI）辅助下进行",
+	"section": "edit",
+	"default": true,
+	"dependencies": ["ext.gadget.i18n", "ext.gadget.Util", "mediawiki.util", "oojs-ui-core"],
+	"rights": ["edit"]
+}
