+++ src/Editform_AiAssisted/modules/constant.ts
@@ -1,0 +1,3 @@
+const INPUT_ID = 'editform_ai_assisted';
+
+export {INPUT_ID};
