+++ src/PreviewWithVariant/definition.json
@@ -2,6 +2,13 @@
 	"enable": true,
 	"description": "<span id=\"Gadget-PreviewWithVariant\"></span> 在源代码编辑器下，以用户指定的语言变体显示预览",
 	"section": "edit",
-	"dependencies": ["ext.gadget.i18n", "mediawiki.Uri", "oojs-ui", "oojs-ui-core", "user.options"],
+	"dependencies": [
+		"ext.gadget.i18n",
+		"mediawiki.jqueryMsg",
+		"mediawiki.Uri",
+		"oojs-ui",
+		"oojs-ui-core",
+		"user.options"
+	],
 	"rights": ["edit"]
 }
