--- src/PreviewWithVariant2017/LICENSE
@@ -1,8 +1,0 @@
-/**
- * SPDX-License-Identifier: MIT
- * _addText: '{{Gadget Header|license=MIT}}'
- *
- * @base {@link https://zh.wikipedia.org/wiki/MediaWiki:Gadget-PreviewWithVariant2017.js}
- * @base {@link https://zh.wikipedia.org/wiki/MediaWiki:Gadget-PreviewWithVariant2017.css}
- * @source {@link https://git.qiuwen.net.cn/InterfaceAdmin/QiuwenGadgets/src/branch/master/src/PreviewWithVariant2017}
- */
